upgpkg: migraphx 5.0.2-1
I was unable to get Python to build at this stage due to an incorrectly
set `INTERFACE_INCLUDE_DIRECTORIES` with pybind11.

Additionally, this appears to be incompatible with the Half library v2,
thus v1 has been added as a source (as it is a single header). Hopefully
this can be revisited later.

Signed-off-by: JP-Ellis <josh@jpellis.me>